+++ __init__.py
@@ -2,7 +2,7 @@ import logging
 import os
 from pathlib import Path
 
-__version__ = "3.3.3"
+__version__ = "3.3.4"
 
 import torch
 from packaging.version import InvalidVersion, Version